Moved setServiceTemplateDefaults() from enterprise to resources, and sharing it across everything that uses CommonSpec, so that spark services benefit from the defaults/validation.
