+++ pkg/splunk/enterprise/configuration_test.go
@@ -411,53 +411,3 @@ func TestSetVolumeDefault(t *testing.T) {
 		}
 	}
 }
-
-func TestSetServiceTemplateDefaults(t *testing.T) {
-	cr := enterprisev1.IndexerCluster{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "stack1",
-			Namespace: "test",
-		},
-		Spec: enterprisev1.IndexerClusterSpec{
-			CommonSplunkSpec: enterprisev1.CommonSplunkSpec{
-				CommonSpec: enterprisev1.CommonSpec{
-					ServiceTemplate: corev1.Service{
-						Spec: corev1.ServiceSpec{
-							Ports: []corev1.ServicePort{
-								{
-									Name: "http",
-									Port: 80,
-								},
-								{
-									Name: "https",
-									Port: 443,
-									TargetPort: intstr.IntOrString{
-										Type:   intstr.Int,
-										IntVal: 8443,
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	setServiceTemplateDefaults(&cr.Spec.CommonSplunkSpec)
-	for _, p := range cr.Spec.CommonSplunkSpec.CommonSpec.ServiceTemplate.Spec.Ports {
-		switch p.Name {
-		case "http":
-			if p.TargetPort.IntValue() != int(p.Port) {
-				t.Errorf("setServiceTemplateDefaults() did not set target port correctly. Want %d, Got %d", p.Port, p.TargetPort.IntVal)
-			}
-			if p.Protocol != corev1.ProtocolTCP {
-				t.Errorf("setServiceTemplateDefaults() did not set protocol correctly. Want %s, Got %s", corev1.ProtocolTCP, p.Protocol)
-			}
-		case "https":
-			if p.TargetPort.IntValue() != 8443 {
-				t.Errorf("setServiceTemplateDefaults() did not set target port correctly. Want 8443, Got %d", p.TargetPort.IntVal)
-			}
-		}
-	}
-}
