+++ pkg/splunk/spark/configuration_test.go
@@ -83,7 +83,7 @@ func TestGetSparkService(t *testing.T) {
 	}
 
 	test(SparkMaster, false, `{"kind":"Service","apiVersion":"v1","metadata":{"name":"splunk-stack1-spark-master-service","namespace":"test","creationTimestamp":null,"labels":{"app.kubernetes.io/component":"spark","app.kubernetes.io/instance":"splunk-stack1-spark-master","app.kubernetes.io/managed-by":"splunk-operator","app.kubernetes.io/name":"spark-master","app.kubernetes.io/part-of":"splunk-stack1-spark"},"ownerReferences":[{"apiVersion":"enterprise.splunk.com/v1alpha3","kind":"Spark","name":"stack1","uid":"05db21b4-7244-4022-a844-c131a8747f30","controller":true}]},"spec":{"ports":[{"name":"sparkmaster","port":7777,"targetPort":0},{"name":"sparkwebui","port":8009,"targetPort":0}],"selector":{"app.kubernetes.io/component":"spark","app.kubernetes.io/instance":"splunk-stack1-spark-master","app.kubernetes.io/managed-by":"splunk-operator","app.kubernetes.io/name":"spark-master","app.kubernetes.io/part-of":"splunk-stack1-spark"}},"status":{"loadBalancer":{}}}`)
-	test(SparkMaster, true, `{"kind":"Service","apiVersion":"v1","metadata":{"name":"splunk-stack1-spark-master-headless","namespace":"test","creationTimestamp":null,"labels":{"app.kubernetes.io/component":"spark","app.kubernetes.io/instance":"splunk-stack1-spark-master","app.kubernetes.io/managed-by":"splunk-operator","app.kubernetes.io/name":"spark-master","app.kubernetes.io/part-of":"splunk-stack1-spark"},"ownerReferences":[{"apiVersion":"enterprise.splunk.com/v1alpha3","kind":"Spark","name":"stack1","uid":"05db21b4-7244-4022-a844-c131a8747f30","controller":true}]},"spec":{"ports":[{"name":"sparkmaster","port":7777,"targetPort":0},{"name":"sparkwebui","port":8009,"targetPort":0}],"selector":{"app.kubernetes.io/component":"spark","app.kubernetes.io/instance":"splunk-stack1-spark-master","app.kubernetes.io/managed-by":"splunk-operator","app.kubernetes.io/name":"spark-master","app.kubernetes.io/part-of":"splunk-stack1-spark"},"clusterIP":"None"},"status":{"loadBalancer":{}}}`)
+	test(SparkMaster, true, `{"kind":"Service","apiVersion":"v1","metadata":{"name":"splunk-stack1-spark-master-headless","namespace":"test","creationTimestamp":null,"labels":{"app.kubernetes.io/component":"spark","app.kubernetes.io/instance":"splunk-stack1-spark-master","app.kubernetes.io/managed-by":"splunk-operator","app.kubernetes.io/name":"spark-master","app.kubernetes.io/part-of":"splunk-stack1-spark"},"ownerReferences":[{"apiVersion":"enterprise.splunk.com/v1alpha3","kind":"Spark","name":"stack1","uid":"05db21b4-7244-4022-a844-c131a8747f30","controller":true}]},"spec":{"ports":[{"name":"sparkmaster","port":7777,"targetPort":0},{"name":"sparkwebui","port":8009,"targetPort":0}],"selector":{"app.kubernetes.io/component":"spark","app.kubernetes.io/instance":"splunk-stack1-spark-master","app.kubernetes.io/managed-by":"splunk-operator","app.kubernetes.io/name":"spark-master","app.kubernetes.io/part-of":"splunk-stack1-spark"},"clusterIP":"None","type":"ClusterIP"},"status":{"loadBalancer":{}}}`)
 
 	cr.Spec.ServiceTemplate.Spec.Type = "LoadBalancer"
 	cr.Spec.ServiceTemplate.ObjectMeta.Labels = map[string]string{"1": "2"}
@@ -91,5 +91,5 @@ func TestGetSparkService(t *testing.T) {
 	cr.ObjectMeta.Annotations = map[string]string{"a": "b"}
 
 	test(SparkWorker, false, `{"kind":"Service","apiVersion":"v1","metadata":{"name":"splunk-stack1-spark-worker-service","namespace":"test","creationTimestamp":null,"labels":{"1":"2","app.kubernetes.io/component":"spark","app.kubernetes.io/instance":"splunk-stack1-spark-worker","app.kubernetes.io/managed-by":"splunk-operator","app.kubernetes.io/name":"spark-worker","app.kubernetes.io/part-of":"splunk-stack1-spark","one":"two"},"annotations":{"a":"b"},"ownerReferences":[{"apiVersion":"enterprise.splunk.com/v1alpha3","kind":"Spark","name":"stack1","uid":"05db21b4-7244-4022-a844-c131a8747f30","controller":true}]},"spec":{"ports":[{"name":"workerwebui","port":7000,"targetPort":0},{"name":"dfwreceivedata","port":17500,"targetPort":0}],"selector":{"app.kubernetes.io/component":"spark","app.kubernetes.io/instance":"splunk-stack1-spark-worker","app.kubernetes.io/managed-by":"splunk-operator","app.kubernetes.io/name":"spark-worker","app.kubernetes.io/part-of":"splunk-stack1-spark"},"type":"LoadBalancer"},"status":{"loadBalancer":{}}}`)
-	test(SparkWorker, true, `{"kind":"Service","apiVersion":"v1","metadata":{"name":"splunk-stack1-spark-worker-headless","namespace":"test","creationTimestamp":null,"labels":{"app.kubernetes.io/component":"spark","app.kubernetes.io/instance":"splunk-stack1-spark-worker","app.kubernetes.io/managed-by":"splunk-operator","app.kubernetes.io/name":"spark-worker","app.kubernetes.io/part-of":"splunk-stack1-spark","one":"two"},"annotations":{"a":"b"},"ownerReferences":[{"apiVersion":"enterprise.splunk.com/v1alpha3","kind":"Spark","name":"stack1","uid":"05db21b4-7244-4022-a844-c131a8747f30","controller":true}]},"spec":{"ports":[{"name":"workerwebui","port":7000,"targetPort":0},{"name":"dfwreceivedata","port":17500,"targetPort":0}],"selector":{"app.kubernetes.io/component":"spark","app.kubernetes.io/instance":"splunk-stack1-spark-worker","app.kubernetes.io/managed-by":"splunk-operator","app.kubernetes.io/name":"spark-worker","app.kubernetes.io/part-of":"splunk-stack1-spark"},"clusterIP":"None"},"status":{"loadBalancer":{}}}`)
+	test(SparkWorker, true, `{"kind":"Service","apiVersion":"v1","metadata":{"name":"splunk-stack1-spark-worker-headless","namespace":"test","creationTimestamp":null,"labels":{"app.kubernetes.io/component":"spark","app.kubernetes.io/instance":"splunk-stack1-spark-worker","app.kubernetes.io/managed-by":"splunk-operator","app.kubernetes.io/name":"spark-worker","app.kubernetes.io/part-of":"splunk-stack1-spark","one":"two"},"annotations":{"a":"b"},"ownerReferences":[{"apiVersion":"enterprise.splunk.com/v1alpha3","kind":"Spark","name":"stack1","uid":"05db21b4-7244-4022-a844-c131a8747f30","controller":true}]},"spec":{"ports":[{"name":"workerwebui","port":7000,"targetPort":0},{"name":"dfwreceivedata","port":17500,"targetPort":0}],"selector":{"app.kubernetes.io/component":"spark","app.kubernetes.io/instance":"splunk-stack1-spark-worker","app.kubernetes.io/managed-by":"splunk-operator","app.kubernetes.io/name":"spark-worker","app.kubernetes.io/part-of":"splunk-stack1-spark"},"clusterIP":"None","type":"ClusterIP"},"status":{"loadBalancer":{}}}`)
 }
